+++ src/map/packets/char_health.cpp
@@ -37,7 +37,7 @@ CCharHealthPacket::CCharHealthPacket(CCharEntity* PChar)
 
 	WBUFL(data,(0x08)-4) = PChar->health.hp;
 	WBUFL(data,(0x0C)-4) = PChar->health.mp;
-	WBUFL(data,(0x10)-4) = (int16)PChar->health.tp;
+	WBUFL(data,(0x10)-4) = PChar->health.tp;
 
 	WBUFW(data,(0x14)-4) = PChar->targid;
